requirements.txt
json
discord.py
flask
python-dotenv
webserver
asyncio
discord.utils
